move csv field helpers into a module
