+++ app/app.rb
@@ -1,7 +1,9 @@
 require 'csv'
+require_relative '../lib/csv_field_helpers'
 
 module OpenDataMaker
   class App < Padrino::Application
+    include CSVFieldHelpers
     register SassInitializer
     register Padrino::Helpers
 
@@ -51,34 +53,6 @@ module OpenDataMaker
       data.to_json
     end
 
-    # Where should these two methods be defined?
-    def recursive_keys(row, prefix = '', path = [])
-      human_names = []
-      paths = []
-      row.keys.each do |key|
-        if row[key].is_a?(Hash)
-          new_human_names, new_paths = recursive_keys(row[key], key + '.', path + [key])
-          human_names += new_human_names
-          paths += new_paths
-        else
-          human_names << prefix + key
-          paths << path + [key]
-        end
-      end
-
-      [human_names, paths]
-    end
-
-    def recursive_fields(row, path_list)
-      path_list.map do |paths|
-        current_value = row
-        paths.each do |path_entry|
-          current_value = current_value[path_entry]
-        end
-        current_value
-      end
-    end
-
     get :index, :with => :endpoint do
       endpoint = params.delete('endpoint')
 
@@ -86,13 +60,11 @@ module OpenDataMaker
       if match_data = format_regex.match(endpoint)
         format = match_data[1]
         endpoint = endpoint.sub(format_regex, '')
+      else
+        format = 'json'
       end
 
-      if format == 'csv'
-        content_type :csv
-      else
-        content_type :json
-      end
+      content_type(format == 'csv' ? :csv : :json)
       headers 'Access-Control-Allow-Origin' => '*',
                'Access-Control-Allow-Methods' => ['GET']
                
@@ -109,18 +81,14 @@ module OpenDataMaker
       sort = params.delete('sort')
       data = DataMagic.search(params, sort:sort, api:endpoint, fields:fields)
       
-      # We'd like to extract this but don't know where to put it.
       if format == 'csv'
         csv_data = data['results']
-        # we're going to assume all rows have the same keys
-        
+        # We assume all rows have the same keys
         if csv_data.first
           human_names, paths = recursive_keys(csv_data.first)
           CSV.generate(force_quotes: true, headers: true) do |csv|
             csv << human_names
-            csv_data.each do |row|
-              csv << recursive_fields(row, paths)
-            end
+            csv_data.each { |row| csv << recursive_fields(row, paths) }
           end
         else
           ''
